Document Play time manager. unk_02017428
