--- include/unk_02017428.h
@@ -1,10 +1,0 @@
-#ifndef POKEPLATINUM_UNK_02017428_H
-#define POKEPLATINUM_UNK_02017428_H
-
-#include "play_time.h"
-
-void PlayTime_FlagNotStarted(void);
-void PlayTime_Start(PlayTime *playTime);
-void sub_02017458(void);
-
-#endif // POKEPLATINUM_UNK_02017428_H
