+++ platinum.us/main.lsf
@@ -65,7 +65,7 @@ Static main
 	Object main.nef.p/src_unk_02015F84.c.o
 	Object main.nef.p/src_unk_02017038.c.o
 	Object main.nef.p/src_easy3d_object.c.o
-	Object main.nef.p/src_unk_02017428.c.o
+	Object main.nef.p/src_play_time_manager.c.o
 	Object main.nef.p/src_unk_02017498.c.o
 	Object main.nef.p/src_system.c.o
 	Object main.nef.p/src_heap.c.o
