+++ src/main.c
@@ -19,6 +19,7 @@
 #include "main.h"
 #include "math_util.h"
 #include "overlay_manager.h"
+#include "play_time_manager.h"
 #include "rtc.h"
 #include "save_player.h"
 #include "savedata.h"
@@ -27,7 +28,6 @@
 #include "sys_task_manager.h"
 #include "system.h"
 #include "touch_pad.h"
-#include "unk_02017428.h"
 #include "unk_02022844.h"
 #include "unk_0202419C.h"
 #include "unk_0202CC64.h"
@@ -144,7 +144,7 @@ void NitroMain(void)
         }
 
         UpdateRTC();
-        sub_02017458();
+        PlayTime_IncrementTimer();
         sub_020241CC();
         SysTaskManager_ExecuteTasks(gSystem.printTaskMgr);
